Move dependency management from yarn to npm (#254)
Fixes #40.

During development of the diagnose function, I noticed that yarn was missing a lot of functionality needed to test nodejs CLI binaries. As this also fixes #40, I have decided to migrate the project from yarn to npm.

This gives us some better developer ergonomics and makes us more able to accept contributions, at the cost of losing the yarn workspaces feature, which we will regain again once npm v7 is out. This means that each package needs a lockfile of it's own going forward