Set Express route params as `params` sample data (#272)
Fixes a recent support query.

This appends the parsed dynamic route parameters (req.params) and query string (req.query, if it exists) from Express to the root span as sample data.

This required a small refactor to the data model - rather than exposing two separate classes for DataMap and DataArray, I have opted to more closely match the Ruby implementation. My thinking at the time was that it might be useful if those objects could be introspect-able, but on reflection, I don't see much need for this.

This also closes appsignal/appsignal-agent#619 by using a filtered map for sample data, allowing data to be filtered by the agent, and adding associated config for this.