Fix musl target and override reporting
The musl_override value in the diagnose reports if the user has manually
chosen the linux-musl build of the agent and extension, by specifying
`APPSIGNAL_BUILD_FOR_MUSL=true` as the environment variable. (Also
accepts the `1` value on other integrations, see
https://github.com/appsignal/integration-guide/pull/26 .)

In this Node.js integration it currently is set to true if the host
system is detected as a musl system, not just when the user specifies
the `APPSIGNAL_BUILD_FOR_MUSL` environment variable.

This change fixes that, and moves the `-musl` suffix to the `target`
key, which is used to determine which file to download. The `-musl`
suffix should be part of the `target` key value, like it's done in other
integrations.

Fixes https://github.com/appsignal/appsignal-nodejs/issues/370

I've not implemented the fallback for musl builds on old libc versions
like is part of the other integrations, as that's outside of the scope
of this change.