Handle all non-dynamic paths as static routes
In previous versions, we only used the full path name as in the sample
name if it matched "/". This patch removes that constraint, leaving two
types of routes:

1. Dyanmic routes, matched from the dynamicRoutes
2. Static routes, including everything else

This is based on the assumption that all non-dynamic routes are in fact
static. I haven't found any examples where this might fail, but I'm open
to suggestions on that.

Closes #337.