Make APPSIGNAL_BUILD_FOR_MUSL behavior consistent (#396)
Other integrations listen to both the `true` and `1` value. We only
documented the `1` value in our docs here:
https://docs.appsignal.com/support/operating-systems.html#musl-build-override

This means that users may try to use the `1` value but it doesn't work
for them. This makes the `1` value work now.

As decided upon in:
https://github.com/appsignal/integration-guide/pull/26

I also added a Jest testing setup. Configure it to reset all mocks and
modules between tests so that state from mocks don't leak between tests.

I mocked some calls to the Node.js `process` global because I needed to
test the results for different platforms and environment variables.

Fixes #368