Add mono to project (#394)
* Add mono to project

Replace lerna with mono, our own tool for managing mono repos.
https://github.com/appsignal/mono/

Follow the instructions on the mono README to install it locally.

Mono is also installed on the CI using the `script/setup` script
which is run before every job. The mono download is cached between
jobs so we don't download it for every job fresh. Update the version
variable `MONO_VERSION` to update the mono install on the CI.
This `setup` script will check the existence of the `cache` command, if
not it will skip it. This allows it to be reused in the test-setups
repo.

Mono depends on npm workspaces, which are available since npm 7. npm 7
also uses lockfileVersion "2" which is a slightly different format, but
causes a lot of unrelated changes in this diff. The `package-lock.json`
files can be removed from the individual packages with workspaces too.

* Update mono to 0.2.0 (#395)

Update mono so we can use the `--package` selector and run the
`build:ext` script through mono. This way we use a consistent interface.

* Update to mono v0.3.0 (#397)

- Update README to mention mono instead of lerna.
- Use --parallel for "build:watch" command.
- Use new setup script location.

* Update README [ci skip]

* Update mono to v.0.4.0

This version updates how the clean command works (it's new and the old
behavior was moved to the "unbootstrap" command) and allow packages to
extend the clean script in the `package.json`.

* Update mono to 0.5.0-alpha.1

Add support for updating dependent packages in Node.js projects.

* Update to mono 0.5.0

* Update to mono 0.5.1

Fix npm publish command in mono repo.

* Update to mono 0.5.2

Remove tag_prefix config option from mono.yml, it has been removed since
it was no longer necessary.