Stub out package.json root scripts
When mono is called from `npm run` scripts it inherits some context in
environment variables, which make it run the script in the root package
again (which is a mono command), causing an infinite loop. Even when
specifying a `--workspace`, like mono does, it will not run the root
package script.

A bunch of environment variables are set by `npm run`, bu the
environment variable PATH is the one that causes this behavior. I don't
know how mono could unset it without knowing the original PATH value.

Stub the commands out for now so users don't run into this issue.