Add build matrix
Add experimental build matrix for Node.js to test on different versions
of Node.js and against different versions of the packages we support
integrations with.

In the `build_matrix.yml` it is specified which packages to test and
against which versions. These are specified as "variations" on the
normal test suite.

Each variation has a packages object of packages (key) and their
versions (value). These will be installed for the jobs that test these
variations using `npm install`. It may be that it's not enough to only
specify the package we integrate with (like next), but also it's other
peer dependencies (like react and react-dom).

The installation of specific package versions will update the
package.json and package-lock.json in the project, but should be
committed.