Add build matrix
Add experimental build matrix for Node.js to test on different versions
of Node.js and against different versions of the packages we support
integrations with.

## The build matrix

In the `build_matrix.yml` it is specified which packages to test and
against which versions. These are specified as "variations" on the
normal test suite.

Each variation has a packages object of packages (key) and their
versions (value). These will be installed for the jobs that test these
variations using `npm install`. It may be that it's not enough to only
specify the package we integrate with (like next), but also it's other
peer dependencies (like react and react-dom).

The installation of specific package versions will update the
package.json and package-lock.json in the project, but should be
committed.

Some packages have other tests that need to run in a different way, such
as next.js, these tests are specified under a special "extra_tests" key.

## Limited builds added

I've only added variations for packages that have multiple versions of
dependencies we can tests against. I've also added a "latest" version
build, so that we can spot failures against newer versions of a package
early, but then we should also add new variations for new releases after
they come out. Now we may run builds twice if one of the specified
version is also the latest version.

## Leaving out builds

I've left out several builds in this implementation because they fail. I
want this setup to first be merged as a valid setup before spending time
on fixing builds.

Builds I did not include because they are currently broken:

- Node.js
- Next.js version 9.5.5

## Implementation in Ruby

I've written the `semaphore.yml` file generation script in Ruby as it
started off with a copy from the Ruby gem, but I modified it for the
Node.js build with its own rules.
https://github.com/appsignal/appsignal-ruby/blob/3e31416039b6570150dbfe2fa6ee85a238efa61b/Rakefile#L46-L104