Return the full log path in the diagnose report (#402)
* Return the full log path in the diagnose report

Currently, the diagnose command expects the logPath to include the
filename for the `appsignal.log` file. To get the `log_dir_path`, the
filename is stripped off. The `LogPath` is then used to return the full
path to the file.

In reality, it should be the other way around. The `LogPath` does not
include the filename, so it can be used as the log_dir_path in the
diagnose. Then, to get the full path to the `appsignal.log` file, this
patch appends the filename to the log path.

* Use internal value for logFilePath in diagnose

Co-authored-by: Tom de Bruijn <tom@tomdebruijn.com>

* Explicitly cast logFilePath to <string>

* Diagnose strips "/appsignal.log" from directory path, if present

Co-authored-by: Tom de Bruijn <tom@tomdebruijn.com>