Bump Next.js version required to version 10.2.3 (#407)
There's a compatibility issue with the 9.5.5 release of Next.js.
If we start the integrations test with Next.js 9.5.5 we get the
following error:

```
/home/semaphore/appsignal-nodejs/packages/nodejs/dist/scope.js:109
            throw err;
            ^
TypeError: Cannot read property 'router' of undefined
    at /home/semaphore/appsignal-nodejs/packages/nextjs/dist/handler.js:16:39
    at Server.<anonymous> (/home/semaphore/appsignal-nodejs/packages/nextjs/test/example/server.js:32:5)
    at Server.emit (events.js:315:20)
    at /home/semaphore/appsignal-nodejs/packages/nodejs/dist/instrumentation/http/lifecycle/incoming.js:67:29
    at ScopeManager.withContext (/home/semaphore/appsignal-nodejs/packages/nodejs/dist/scope.js:105:20)
    at BaseTracer.withSpan (/home/semaphore/appsignal-nodejs/packages/nodejs/dist/tracer.js:48:68)
    at Server.<anonymous> (/home/semaphore/appsignal-nodejs/packages/nodejs/dist/instrumentation/http/lifecycle/incoming.js:54:23)
    at parserOnIncoming (_http_server.js:779:12)
    at HTTPParser.parserOnHeadersComplete (_http_common.js:117:17)
```

As discussed, we will not fix this issue. Instead spending more time
adding tests for other integrations.

To clarify that we do not support Next.js 9.x, increase the required
version to the latest 10.x release that does work.