Fix extension installation reporting
When the extension installed we never stored if it installed correctly
or not. Originally it would always report it as a "success" status. In a
previous commit cb08b7a16c4f81239724a1f71ac99371015b568e, PR #373, I
changed the report status to "unknown", because after the `extension.js`
script was run the `node-gyp rebuild` command was run through the
`package.json` `install` script. After the node-gyp command was run it
would not return to the extension.js script and we couldn't check the
exit status of that command.

Instead of running the node-gyp command after the `extension.js` script,
run the node-gyp command from the `extension.js` script so we can listen
to the exit status and report the installation as a success or "error"
when it encounters an error.

This should fix #371, the issue describing the scenario above, the
install report not accurately describing the installation status.

I am not sure why the originally implementation was chosen. The original
implementation was present since the initial commit of the project.
There may be a problem with this new implementation that I am unaware of
now, because I'm no Node.js expert. But we call out to other commands
like `tar` from the `extension.js` script as well, so I figure it will
be fine.

## CI build update

I've added tests for the extension build failure scenario, and the
success scenario. For this I needed to have Jest run certain specs only
some of the time, based on the `_TEST_APPSIGNAL_EXTENSION_FAILURE`
environment variable (same as in the Ruby gem CI setup). This is done
with the `--filter` option (which links to a file with a filter
function), which is configured through the `package.json` test scripts.

Since these new tests are only for the `nodejs-ext` package I found it
unnecessary to spin up a new job for it, and instead add it as
`extra_commands` to the package's test commands.

Before the failure state is tested, the extension installation is
cleaned up, so it doesn't think it has installed correctly.

Because we install the extension in the "Build" task, and it's not
present in the project dir, but instead the `/tmp` dir on the system,
I've also copied the report location from the "Build" task to the job
that tests this report. Otherwise it would fail on a missing file, or
we'd need to install the extension again. Hopefully we can find a better
location for the install report location in issue #372.