Write install report to a unique path for install (#412)
The installation report was written to
`/tmp/appsignal-install-report.json`, which can be easily overwritten by
other installations of the AppSignal nodejs-ext package in other apps.

To avoid reports being overwritten, and only the last report being
available, write the install report to a path unique to the app install.

This change generates a digest for the app's location that's used by the
`extension.js` script to write to and the `diagnose.ts` tool to read
from. This path should match as long as the app hasn't been moved after
installation. If it doesn't match the diagnose report won't be able to
find the installation report.

I've implemented the `reportPath` function twice, in both packages,
because they can't share code as the nodejs-ext package may not have
installed correctly and been removed. The implementation of these two
functions need to be kept in sync in order to generate the same path
digest.

## Alternatives

It's not possible currently to write the install report in the
`nodejs-ext` package itself, like the `ext/` dir where we also download
the extension and agent files. If the package fails to install
its directory is removed after the `npm install` process, so we need to
store it somewhere outside of the package directory.

An alternative solution would be to not exit the install process with an
error status code. This way the extension package directory isn't
removed, but we've decided not to do that for now while discussing
issue #372, as it could have impact on the compilation process we're not
foreseeing.

Closes #372