Fix diagnose report logging 10 appsignal.log lines
When the appsignal.log file is read and split into an array of lines,
the last line of the log, which is always empty, is also considered a
line to be printed.

This is because when a String is split on a line break, it will return
that last empty line as a empty String.

```
$ node
> "foo\nbar\nbaz\n".split("\n")
[ 'foo', 'bar', 'baz', '' ]
```

Since it's empty it shows no useful information and
breaks the formatting, so trim the read data to not include that last
empty line.