Fix Push API Key validation in diagnose report
The Push API key validation wasn't implemented previously. I've now
implemented a validation by calling the AppSignal.com auth endpoint with
the configured Push API key.

The http(s) library doesn't a way to make synchronous requests that I
could find, so I had to update the diagnose tool and script to be async,
by using the async and await keywords all the way up the chain of
functions.

The rest of the implementation such as the coloring of the output, and
storing of the result on the report, was already implemented.

I've updated the diagnose tests submodule to account for this change in
behavior in the tests.

Fix #408