Remove unused test jobs on CI
For some reason the Node.js test jobs were generated on Jeff's machine,
but not on mine. Most likely something on Jeff's machine exists that is
recognized as a test dir/file, but is not checked in.

There are no test files/directories checked in, so remove the Node.js
test jobs from the Semaphore config.