Print skipped package tests in build matrix
This way we can see when package tests are omitted from the
`.semaphore.yml` file. I don't expect anyone to review the many line
changes in that file all the time. Now you can see what happens in the
command output.

Make it so that it only logs every package once, and not for every
combination of Node.js version as well.