Test non-root path in Express tests
Do not only test the root path, but test another path that can exist in
the app. This is how I found out that the AppSignal middleware that was
required was not actually used in the test app.

Without the middleware it would set the action name to:
`GET [unknown route]`