Add helper script to run integration tests
I've added a convenience helper script to run the integration tests more
easily. Call the script with the package directory as an executable and
it will run the integration tests. No need to set environment variables
and remember/copy-paste the command.

I've also updated the commands for the CI build to use this script as
well.

## Alternative implementation

Rather than set the `BUNDLE_GEMFILE` env variable and specify the path
to run the tests, navigate to the directory with the test suite and run
the test suite from that context. This requires less configuration of
the bundle gemfile and path in which the tests are present.

Instead of always running `bundle install`, first run `bundle check` if
it's really necessary. This will print less output per test run.