Refactor integration tests to reuse helpers
Add an AppRunner class with all the logic to start and stop the example
app. Since it knows a bit about the app, we can also ask it about the
output and the log contents.

With this AppRunner all the behavior is put in one place and not all the
integration specs need to reimplement this.

Remove the usage of `Dir.chdir` which changes the entire Ruby app's
current working directory and pass along the `chdir` option to the
`spawn` command that starts the app so that only the example app is run
in that directory.