Add RuboCop linter for Ruby code
Parts of the repository contains Ruby code. Let's add RuboCop to it to
ensure we follow a common code style.

I let RuboCop fix most of the issues, but things like line length had to
be fixed manually.

I modernized the Ruby gem's RuboCop config (as it's locked on an older
version of RuboCop) and made it consistent with the RuboCop config in
the diagnose tests repo.

https://github.com/appsignal/diagnose_tests/pull/6