Use fetch_root_span_id helper in integration specs
Add a `fetch_root_span_id` helper that will extract the root span id
from the logs. This mean we won't have to repeat that logic for every
spec or every spec file. It's a common pattern that we do not need to
repeat.

It doesn't match the namespace name now. If we need to match that we can
also add an option to the helper that adds the namespace to the regular
expression.

The log line match we do using the span id doesn't need to be a regular
express, we only match a plain string, remove the regular express and
use `String#include` instead.