Fix test detection for build matrix
If there's an example app for a package integration test in the
directory, it will also scan the `node_modules` directory for tests. So
even if the package doesn't have tests, if a dependency has tests, it's
recognized as having tests.

Filter out files from the `node_modules` to not recognize them as tests
for the package, and not generate jobs for those packages.