Update integration apps dependencies
Do not lock the packages in the `package.json` file because they will be
updated on the CI to test against different versions. Change the version
specified to an asterisk `*` to indicate we don't care about the version
locally.

Remove the `package-lock.json` files (for now), as it locks packages
and their dependencies to certain versions, which for some reason
breaks sometimes when we up or downgrade packages in the build matrix
on the CI.