Start integration test apps with npm
By starting the app using `node file.js` we do not rely on the packages
installed by npm for that specific package. It looks like it will search
the global installed packages and/or navigate up the tree until it finds
a `node_modules` where it can load the packages from.

Using `npm run start` we start the app in the npm context as specified
by the `package.json` for the example app. This way it should load the
version of the dependencies as specified in `package.json`.

Because we start the app through an npm script the handling of stopping
the app is slightly different. We can't reliably find the app process by
name as it's wrapped by npm as the "(node)" process.

Sending the kill process to the npm wrapper doesn't always work. On
macOS we need to send signal 2 and on the CI we need to send signal 3. I
do not know at this time why this is different. If we don't send these
signals the app will not exit and the test suite will fail to exit as it
cannot shutdown the example app before stopping. This is especially
problematic locally as it makes running the test suite multiple times
very difficult and the user needs to exit the app manually after every
run.