Add root span helper in tracer
The actual root span wasn't accessible as there were no helpers nor sets
for it. Making adding errors a difficult or impossible task if tried
  from nested span contexts.

This commit allows access to the root span from everywhere through the
tracer using similar storage as scopes, so we keep taking advantage of
Async Hooks to make it safe for async contexts.