Add error tracking to RootSpan
Add a new helper function to Tracer so we track errors properly into the
root span.

It also includes the first step of the cleanup of the interfaces
starting from the Client and the Tracer.

Add a new helper function to Tracer to track errors properly into the
root span.