Print install report on job failure
When the build or test jobs fail, print the extension installation
report. Something may have gone wrong with the installation and being
able to view that right on the job page is very useful. Otherwise we
would need to SSH to the job and read it from the filesystem.

I've added it to the build matrix as a custom property "epilogue"
that's copied to all build and test jobs, so it's easy to customize
it later.