Improve install report generation
Only set the download report once, before it's downloaded and update it
along the different steps in the process. Only set if the download was
verified when it was verified, and default to "unverified".

This way we don't have to generate the report twice (success and failure
state), and it's more accurate in reflecting the state throughout the
process, and when it fails during one of the steps.