Set build source location in the install report
The install report did not track which source was used during
installation. Currently the Node.js install report only supports the
remote source, but I want to update it to also write reports for local
installs later on.

This source key is supported already by the other integrations and this
only updates the Node.js integration to match.