Support local extension files for installer
When we use local agent and extension files to install and test the
local builds, also run it through the install script. This way we can
just use `npm install` rather than having to run the `npm run build:ext`
script in the `nodejs-ext` package.

Previously the install script would immediately exit on detecting local
files. This is unlike other integrations (Ruby and Elixir). It doesn't
store a install report file, so there's no way of knowing if the
extension was installed correctly (when using the `npm run build:ext`
task to manually install the extension).

With this change there's just one way to install the extension, from
local and remote sources. This will help us debug issues where a local
build is used.

When users install the page from npmjs.org they will not notice a
different. If they run `npm install` the extension is only installed the
first time and not on subsequent `npm install`s.

With this change the extension will be installed on every `npm install`
if the packages are linked using `npm link` locally. This is similar to
how it works in Elixir when specifying the AppSignal package in
`mix.exs`. Users won't generally use `npm link` so this is not a
problem.

Closes #437
Reference: https://github.com/appsignal/integration-guide/issues/45