Add rootSpan and addError helpers
Root span wasn't accessible when working with contexts, only the current
child span was. Now there's a different storage for root spans and
they're always accessible from the tracer object.

A new `addError` function has been added to the tracer object and this
is the one that must be used when sending errors to AppSignal. It'll
look for the root span no matter the context or the current span, and
will attach the errors to it so they don't get ignored by the agent.