Deprecate addError functions
In PR #446 we removed the `addError` helper and renamed it `setError`
without leaving in place a placeholder function for `addError` that
prints a deprecation warning. Without this placeholder function users
could run into an error when calling a function that does not exist.

I've readded the `addError` function and made it print a deprecation
warning. I've updated the `setError` function to print the same
deprecation warning.

The types package has been updated to include the `addError` function
again on the Span interface.