Read extension install status from report
The installation result was hardcoded to always be `success`, because
previously we couldn't track it. But since PR #409 we now run the
`node-gyp` command from the `extension.js` install script so we can
track the installation result.

Read the status from the report so the diagnose report is more accurate.

Since there are more fields in case of an error or failure, print those
as well if present.