Read language version from installation report (#452)
In the diagnose report, in the extension installation report section,
the wrong language version field was read. It was using the one from the
"host" section, which is generated upon execution of the diagnose
report.

We want to display the field from the installation report, because that
may differ, and if it does, point to a different installation machine
than the one the app runs on. This could cause problems and we want to
highlight those differences.

This does affect the report sent to AppSignal.com. Only the report
displayed in the terminal.