Diagnose print parsing error on missing report
When the diagnose command is run and the installation report is missing
the diagnose command crashes with an error.

```
UnhandledPromiseRejectionWarning: TypeError: Cannot read property
  'download_url' of undefined
```

This is because when the report can't be read the
`getInstallationReport` function would only return an empty object,
rather than the expected object with a parsing error. (Expected in the
sense that's how our other integrations handle this type of error and
that's what the diagnose tool on the server side expects.)

With this change the report should no longer crash and the sent report
includes the error about the missing installation report.

I didn't know how to resolve a TypeScript issue with the optional `raw`
key on the parsing error object, so I created a ParsingError type to
indicate that it's a optional key. I couldn't get it to work otherwise.

Part of #448