Print diagnose String values surrounded by quotes
This makes the Node.js integration print the diagnose output
consistently with the Ruby integration: String values are surrounded by
quotes when printed.

Move the `format_value` out of the Diagnose class so we don't need to
prefix it with `this.` every time. The function doesn't rely on
anything in the class in terms of state so we can move it out.