Print agent diagnose report in CLI output
The agent diagnose report wasn't printed in the diagnose report as the
only missing section. It was sent with the diagnose report, just not
printed.

I don't understand much of TypeScript. I had to add a lot more type
definitions than I expected for the functions, which is why the new
types exist at the top of the `cli/diagnose.ts` file.

I considered changing the structure of how agent returns this report,
but I was already far enough done a rabbit hole trying to make this work
in the first place.

Looking back I structured the report by the agent and extension in such
as way that it was future proof (adding more tests to it later), but I
haven't found a way to easily parse or format this in any integration or
on the server side validator. This may be a nice thing to change to
make this whole thing more easy to maintain. To change it to an easier
(less nested) format.