Add Node 17 to the build matrix
Modify the build matrix to add Node 17 to the environments under
test. The GCC setup from Node 16 is reused here.

As a workaround for a bug caused by Node 17 switching to
OpenSSL 3, we add the `--openssl-legacy-provider` option to the
`NODE_OPTIONS` environment variable. See [this node issue][node]
and [this webpack issue][webpack] for details.

[node]: https://github.com/nodejs/node/issues/40455
[webpack]: https://github.com/webpack/webpack/issues/14532