Replace type definitions with generic HashMap
The agent diagnose tests interfaces defined a type that's only used in
the diagnose tool itself. This was because I couldn't get the TypeScript
compiler happy any other way.

But really, these types serve no purpose outside of this file and the
code reading from the values with this type is written very defensively,
it should not fail on a missing key or value.
If no value is found it will print a placeholder `-` instead indicating
the value was not present.

Luismi showed me that we can use the `HashMap` type here from the types
package and define the HashMap value as `any` to achieve the same
result.