Allow configuration of diagnose endpoint
For the diagnose tests we need to send the diagnose JSON report to a
custom endpoint. This change introduces the APPSIGNAL_DIAGNOSE_ENDPOINT
environment variable for this purpose.

The APPSIGNAL_DIAGNOSE_ENDPOINT environment variable is not a variable
we'll document publicly, because it's only meant for testing purposes.

The Node.js app needs to either use the https or http module, for the
production or local server. I didn't find a way to call `http:` protocol
endpoints with the https module.