Update the FileMetadata mode type to String
The Ruby gem transmits this value as a String, because I think we
decided to this to make sure we transmit the value in a format we expect
to be human readable, like `00777` or `100777`. It's more flexible as a
String format.

Previously the value would be send as `33188`, but transformed it sends
it as `"100644"`.